Enable admin to delete an app
- marks app for deletion so it can be recovered within 30 days
- admin cannot create app with same name until app gets deleted 30 days after delete is called
- all endpoints that check if app id is valid now also make sure app isn't deleted
