+++ src/userbase-server/user.js
@@ -7,6 +7,7 @@ import crypto from './crypto'
 import connections from './ws'
 import logger from './logger'
 import { validateEmail, stringToArrayBuffer } from './utils'
+import appController from './app'
 
 const getTtl = secondsToLive => Math.floor(Date.now() / 1000) + secondsToLive
 
@@ -50,35 +51,6 @@ const createSession = async function (userId, appId) {
   await ddbClient.put(params).promise()
 
   return { sessionId, creationDate }
-}
-
-const getAppByAppId = async function (appId) {
-  const params = {
-    TableName: setup.appsTableName,
-    IndexName: setup.appIdIndex,
-    KeyConditionExpression: '#appId = :appId',
-    ExpressionAttributeNames: {
-      '#appId': 'app-id'
-    },
-    ExpressionAttributeValues: {
-      ':appId': appId
-    },
-    Select: 'ALL_ATTRIBUTES'
-  }
-
-  const ddbClient = connection.ddbClient()
-  const appResponse = await ddbClient.query(params).promise()
-
-  if (!appResponse || appResponse.Items.length === 0) return null
-
-  if (appResponse.Items.length > 1) {
-    // too sensitive not to throw here. This should never happen
-    const errorMsg = `Too many apps found with app id ${appId}`
-    logger.fatal(errorMsg)
-    throw new Error(errorMsg)
-  }
-
-  return appResponse.Items[0]
 }
 
 const _buildSignUpParams = async (username, passwordSecureHash, appId, userId,
@@ -228,8 +200,8 @@ exports.signUp = async function (req, res) {
     const userId = uuidv4()
 
     // Warning: uses secondary index here. It's possible index won't be up to date and this fails
-    const app = await getAppByAppId(appId)
-    if (!app) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
+    const app = await appController.getAppByAppId(appId)
+    if (!app || app['deleted']) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
 
     const salts = { encryptionKeySalt, dhKeySalt, hmacKeySalt }
     const passwordBasedBackup = { pbkdfKeySalt, passwordEncryptedSeed }
@@ -294,11 +266,11 @@ exports.authenticateUser = async function (req, res, next) {
     // Warning: uses secondary indexes here. It's possible index won't be up to date and this fails
     const [user, app] = await Promise.all([
       getUserByUserId(session['user-id']),
-      getAppByAppId(session['app-id'])
+      appController.getAppByAppId(session['app-id'])
     ])
 
     if (!user || user['deleted']) return res.status(statusCodes['Unauthorized']).send('Session invalid')
-    if (!app) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
+    if (!app || app['deleted']) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
 
     res.locals.user = user // makes user object available in next route
     next()
@@ -406,8 +378,8 @@ exports.signIn = async function (req, res) {
 
   try {
     // Warning: uses secondary index here. It's possible index won't be up to date and this fails
-    const app = await getAppByAppId(appId)
-    if (!app) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
+    const app = await appController.getAppByAppId(appId)
+    if (!app || app['deleted']) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
 
     const params = {
       TableName: setup.usersTableName,
@@ -758,8 +730,8 @@ exports.forgotPassword = async function (req, res) {
 
   try {
     // Warning: uses secondary index here. It's possible index won't be up to date and this fails
-    const app = await getAppByAppId(appId)
-    if (!app) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
+    const app = await appController.getAppByAppId(appId)
+    if (!app || app['deleted']) return res.status(statusCodes['Unauthorized']).send('App ID not valid')
 
     const tempPassword = crypto
       .randomBytes(ACCEPTABLE_RANDOM_BYTES_FOR_SAFE_SESSION_ID)
